chore: Moved io module into common
