+++ ra2ce/analyses/indirect/origin_closest_destination.py
@@ -31,7 +31,7 @@ import pandas as pd
 from shapely.geometry import LineString, MultiLineString
 from tqdm import tqdm
 
-from ra2ce.io.readers.graph_pickle_reader import GraphPickleReader
+from ra2ce.common.io.readers.graph_pickle_reader import GraphPickleReader
 
 
 class OriginClosestDestination:
