+++ ra2ce/common/io/README.md
@@ -1,6 +1,6 @@
 # Purpose of this module.
 
-This module has been created to define generic protocols for read / write functionality as well as to implement concrete generic readers that can be later on reused by specific modules.
+This module has been created to define generic protocols for read / write functionality as well as to implement generic readers that **are used** by different modules.
 
 ## Do not:
 * Create here readers / writers that are very specific for a given module. If only one concrete class is using it, think twice if it should go here.
