+++ ra2ce/common/io/readers/__init__.py
@@ -1,0 +1,1 @@
+from ra2ce.common.io.readers.graph_pickle_reader import GraphPickleReader
