+++ ra2ce/graph/network_config_wrapper.py
@@ -35,7 +35,7 @@ from ra2ce.graph.network_config_data.network_config_data_validator import (
     NetworkConfigDataValidator,
 )
 from ra2ce.graph.networks import Network
-from ra2ce.io.readers import GraphPickleReader
+from ra2ce.common.io.readers import GraphPickleReader
 
 
 def network_handler(config: NetworkConfigData, files: dict) -> Optional[dict]:
