--- ra2ce/io/readers/__init__.py
@@ -1,1 +1,0 @@
-from ra2ce.io.readers.graph_pickle_reader import GraphPickleReader
