+++ tests/graph/exporters/test_json_exporter.py
@@ -3,7 +3,7 @@ import shutil
 import pytest
 
 from ra2ce.graph.exporters.json_exporter import JsonExporter
-from ra2ce.io.writers.ra2ce_exporter_protocol import Ra2ceExporterProtocol
+from ra2ce.common.io.writers.ra2ce_exporter_protocol import Ra2ceExporterProtocol
 from tests import test_results
